refactor: extract isFull and checkIsFull
